# Conflicts:
#   FRONT/muscle_maker/src/app/(beforeLogin)/_component/Main.tsx

Former-commit-id: 0f1033f2aedbb1969f92dfa4fc0cea784702d683

+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/repository/CalendarRepository.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
-public interface CalendarRepository extends JpaRepository<Calendar, Long> {
+public interface CalendarRepository extends JpaRepository<Calendar, Integer> {
     Calendar findCalendarByDateAndClubAndFlagFalse(LocalDate date, Club club);
 
 }
